Create README.md
I'm asking for help with HiJack This because I have all the symptoms of a hijacker connecting and controlling my computer, particularly though XPlane flight simulator. The person(s) have corrupted an instrument on a particular plane that was working before, and tech support at that company is as stumped as I am. This is just the beginning of a series of minor annoyances. I have Zone Alarm, which helps me a lot, but I need more help with this, please. Thanks for your time!